refactor: merge report package into study/report
