+++ cli/study.go
@@ -26,9 +26,9 @@ import (
 
 	"github.com/penny-vault/pvbt/data"
 	"github.com/penny-vault/pvbt/engine"
-	"github.com/penny-vault/pvbt/report"
 	"github.com/penny-vault/pvbt/study"
 	"github.com/penny-vault/pvbt/study/optimize"
+	"github.com/penny-vault/pvbt/study/report"
 	"github.com/penny-vault/pvbt/study/stress"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
