+++ study/integration_test.go
@@ -29,9 +29,9 @@ import (
 	"github.com/penny-vault/pvbt/data"
 	"github.com/penny-vault/pvbt/engine"
 	"github.com/penny-vault/pvbt/portfolio"
-	"github.com/penny-vault/pvbt/report"
 	"github.com/penny-vault/pvbt/study"
 	"github.com/penny-vault/pvbt/study/montecarlo"
+	"github.com/penny-vault/pvbt/study/report"
 	"github.com/penny-vault/pvbt/study/stress"
 )
