+++ study/montecarlo/montecarlo.go
@@ -23,8 +23,8 @@ import (
 
 	"github.com/penny-vault/pvbt/data"
 	"github.com/penny-vault/pvbt/engine"
-	"github.com/penny-vault/pvbt/report"
 	"github.com/penny-vault/pvbt/study"
+	"github.com/penny-vault/pvbt/study/report"
 	"github.com/rs/zerolog/log"
 )
