+++ study/optimize/analyze_test.go
@@ -24,9 +24,9 @@ import (
 	"github.com/penny-vault/pvbt/asset"
 	"github.com/penny-vault/pvbt/data"
 	"github.com/penny-vault/pvbt/portfolio"
-	"github.com/penny-vault/pvbt/report"
 	"github.com/penny-vault/pvbt/study"
 	"github.com/penny-vault/pvbt/study/optimize"
+	"github.com/penny-vault/pvbt/study/report"
 )
 
 // priceAsset is used to build single-row DataFrames for UpdatePrices.
