+++ study/report/format.go
