+++ study/report/metric_pairs.go
