+++ study/report/table.go
