+++ study/report/text.go
