+++ study/report/time_series.go
