+++ study/runner_test.go
@@ -25,8 +25,8 @@ import (
 
 	"github.com/penny-vault/pvbt/engine"
 	"github.com/penny-vault/pvbt/portfolio"
-	"github.com/penny-vault/pvbt/report"
 	"github.com/penny-vault/pvbt/study"
+	"github.com/penny-vault/pvbt/study/report"
 )
 
 // mockStudy implements study.Study for testing the Runner without a real engine.
